+++ app/shared/app-data/src/commonMain/kotlin/data/persistent/SettingsStore.kt
@@ -14,10 +14,10 @@ import androidx.datastore.core.DataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import kotlinx.serialization.builtins.ListSerializer
 import me.him188.ani.app.data.models.danmaku.DanmakuRegexFilter
-import me.him188.ani.app.data.repository.EpisodeHistories
-import me.him188.ani.app.data.repository.MediaSourceSaves
-import me.him188.ani.app.data.repository.MediaSourceSubscriptionsSaveData
-import me.him188.ani.app.data.repository.MikanIndexes
+import me.him188.ani.app.data.repository.media.MediaSourceSaves
+import me.him188.ani.app.data.repository.media.MediaSourceSubscriptionsSaveData
+import me.him188.ani.app.data.repository.media.MikanIndexes
+import me.him188.ani.app.data.repository.player.EpisodeHistories
 import me.him188.ani.utils.io.SystemPath
 
 // 一个对象, 可都写到 common 里, 不用每个 store 都 expect/actual
