+++ app/shared/app-data/src/commonMain/kotlin/data/repository/media/MediaSourceInstanceRepository.kt
@@ -7,12 +7,13 @@
  * https://github.com/open-ani/ani/blob/main/LICENSE
  */
 
-package me.him188.ani.app.data.repository
+package me.him188.ani.app.data.repository.media
 
 import androidx.datastore.core.DataStore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlinx.serialization.Serializable
+import me.him188.ani.app.data.repository.Repository
 import me.him188.ani.app.domain.mediasource.instance.MediaSourceSave
 import me.him188.ani.datasources.api.source.FactoryId
 import me.him188.ani.datasources.api.source.MediaSourceConfig
