+++ app/shared/app-data/src/commonMain/kotlin/data/repository/media/MediaSourceSubscriptionRepository.kt
@@ -7,11 +7,12 @@
  * https://github.com/open-ani/ani/blob/main/LICENSE
  */
 
-package me.him188.ani.app.data.repository
+package me.him188.ani.app.data.repository.media
 
 import androidx.datastore.core.DataStore
 import kotlinx.coroutines.flow.map
 import kotlinx.serialization.Serializable
+import me.him188.ani.app.data.repository.Repository
 import me.him188.ani.app.domain.mediasource.subscription.MediaSourceSubscription
 import me.him188.ani.utils.platform.Uuid
