+++ app/shared/app-data/src/commonMain/kotlin/data/repository/media/MikanIndexCacheRepository.kt
@@ -1,9 +1,19 @@
-package me.him188.ani.app.data.repository
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
+package me.him188.ani.app.data.repository.media
 
 import androidx.datastore.core.DataStore
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 import kotlinx.serialization.Serializable
+import me.him188.ani.app.data.repository.Repository
 import me.him188.ani.datasources.mikan.MikanIndexCacheProvider
 
 interface MikanIndexCacheRepository : Repository, MikanIndexCacheProvider {
