+++ app/shared/app-data/src/commonMain/kotlin/data/repository/player/DanmakuRegexFilterRepository.kt
@@ -7,11 +7,12 @@
  * https://github.com/open-ani/ani/blob/main/LICENSE
  */
 
-package me.him188.ani.app.data.repository
+package me.him188.ani.app.data.repository.player
 
 import androidx.datastore.core.DataStore
 import kotlinx.coroutines.flow.Flow
 import me.him188.ani.app.data.models.danmaku.DanmakuRegexFilter
+import me.him188.ani.app.data.repository.Repository
 
 interface DanmakuRegexFilterRepository : Repository {
