+++ app/shared/app-data/src/commonMain/kotlin/data/repository/player/EpisodeScreenshotRepository.kt
@@ -1,4 +1,13 @@
-package me.him188.ani.app.data.repository
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
+package me.him188.ani.app.data.repository.player
 
 import io.ktor.client.call.body
 import io.ktor.client.request.get
@@ -6,6 +15,7 @@ import io.ktor.client.request.parameter
 import kotlinx.serialization.Serializable
 import me.him188.ani.app.data.models.ApiResponse
 import me.him188.ani.app.data.models.runApiRequest
+import me.him188.ani.app.data.repository.Repository
 import me.him188.ani.utils.ktor.createDefaultHttpClient
 
 interface EpisodeScreenshotRepository : Repository {
