+++ app/shared/app-data/src/commonMain/kotlin/domain/session/BangumiSessionManager.kt
@@ -34,11 +34,11 @@ import me.him188.ani.app.data.models.flatMap
 import me.him188.ani.app.data.models.fold
 import me.him188.ani.app.data.models.map
 import me.him188.ani.app.data.network.BangumiProfileService
-import me.him188.ani.app.data.repository.AccessTokenSession
-import me.him188.ani.app.data.repository.GuestSession
-import me.him188.ani.app.data.repository.Session
-import me.him188.ani.app.data.repository.TokenRepository
-import me.him188.ani.app.data.repository.isValid
+import me.him188.ani.app.data.repository.user.AccessTokenSession
+import me.him188.ani.app.data.repository.user.GuestSession
+import me.him188.ani.app.data.repository.user.Session
+import me.him188.ani.app.data.repository.user.TokenRepository
+import me.him188.ani.app.data.repository.user.isValid
 import me.him188.ani.app.tools.MonoTasker
 import me.him188.ani.app.ui.foundation.BackgroundScope
 import me.him188.ani.app.ui.foundation.HasBackgroundScope
