+++ app/shared/app-data/src/desktopMain/kotlin/domain/media/resolver/DesktopWebVideoSourceResolver.kt
@@ -22,7 +22,7 @@ import kotlinx.serialization.json.Json
 import me.him188.ani.app.data.models.preference.ProxyConfig
 import me.him188.ani.app.data.models.preference.VideoResolverSettings
 import me.him188.ani.app.data.models.preference.configIfEnabledOrNull
-import me.him188.ani.app.data.repository.SettingsRepository
+import me.him188.ani.app.data.repository.user.SettingsRepository
 import me.him188.ani.app.domain.media.resolver.WebViewVideoExtractor.Instruction
 import me.him188.ani.app.platform.AniCefApp
 import me.him188.ani.app.platform.Context
