+++ app/shared/application/src/commonMain/kotlin/platform/CommonKoinModule.kt
@@ -41,30 +41,30 @@ import me.him188.ani.app.data.network.TrendsRepository
 import me.him188.ani.app.data.persistent.dataStores
 import me.him188.ani.app.data.persistent.database.AniDatabase
 import me.him188.ani.app.data.persistent.database.createDatabaseBuilder
-import me.him188.ani.app.data.repository.DanmakuRegexFilterRepository
-import me.him188.ani.app.data.repository.DanmakuRegexFilterRepositoryImpl
-import me.him188.ani.app.data.repository.EpisodeCollectionRepository
-import me.him188.ani.app.data.repository.EpisodePlayHistoryRepository
-import me.him188.ani.app.data.repository.EpisodePlayHistoryRepositoryImpl
-import me.him188.ani.app.data.repository.EpisodePreferencesRepository
-import me.him188.ani.app.data.repository.EpisodePreferencesRepositoryImpl
-import me.him188.ani.app.data.repository.EpisodeProgressRepository
-import me.him188.ani.app.data.repository.EpisodeScreenshotRepository
-import me.him188.ani.app.data.repository.MediaSourceInstanceRepository
-import me.him188.ani.app.data.repository.MediaSourceInstanceRepositoryImpl
-import me.him188.ani.app.data.repository.MediaSourceSubscriptionRepository
-import me.him188.ani.app.data.repository.MikanIndexCacheRepository
-import me.him188.ani.app.data.repository.MikanIndexCacheRepositoryImpl
-import me.him188.ani.app.data.repository.PreferencesRepositoryImpl
 import me.him188.ani.app.data.repository.RepositoryUsernameProvider
-import me.him188.ani.app.data.repository.SettingsRepository
-import me.him188.ani.app.data.repository.SubjectCollectionRepository
-import me.him188.ani.app.data.repository.SubjectSearchHistoryRepository
-import me.him188.ani.app.data.repository.SubjectSearchHistoryRepositoryImpl
-import me.him188.ani.app.data.repository.SubjectSearchRepository
-import me.him188.ani.app.data.repository.TokenRepository
-import me.him188.ani.app.data.repository.TokenRepositoryImpl
-import me.him188.ani.app.data.repository.WhatslinkEpisodeScreenshotRepository
+import me.him188.ani.app.data.repository.episode.EpisodeCollectionRepository
+import me.him188.ani.app.data.repository.episode.EpisodeProgressRepository
+import me.him188.ani.app.data.repository.media.EpisodePreferencesRepository
+import me.him188.ani.app.data.repository.media.EpisodePreferencesRepositoryImpl
+import me.him188.ani.app.data.repository.media.MediaSourceInstanceRepository
+import me.him188.ani.app.data.repository.media.MediaSourceInstanceRepositoryImpl
+import me.him188.ani.app.data.repository.media.MediaSourceSubscriptionRepository
+import me.him188.ani.app.data.repository.media.MikanIndexCacheRepository
+import me.him188.ani.app.data.repository.media.MikanIndexCacheRepositoryImpl
+import me.him188.ani.app.data.repository.player.DanmakuRegexFilterRepository
+import me.him188.ani.app.data.repository.player.DanmakuRegexFilterRepositoryImpl
+import me.him188.ani.app.data.repository.player.EpisodePlayHistoryRepository
+import me.him188.ani.app.data.repository.player.EpisodePlayHistoryRepositoryImpl
+import me.him188.ani.app.data.repository.player.EpisodeScreenshotRepository
+import me.him188.ani.app.data.repository.player.WhatslinkEpisodeScreenshotRepository
+import me.him188.ani.app.data.repository.subject.SubjectCollectionRepository
+import me.him188.ani.app.data.repository.subject.SubjectSearchHistoryRepository
+import me.him188.ani.app.data.repository.subject.SubjectSearchHistoryRepositoryImpl
+import me.him188.ani.app.data.repository.subject.SubjectSearchRepository
+import me.him188.ani.app.data.repository.user.PreferencesRepositoryImpl
+import me.him188.ani.app.data.repository.user.SettingsRepository
+import me.him188.ani.app.data.repository.user.TokenRepository
+import me.him188.ani.app.data.repository.user.TokenRepositoryImpl
 import me.him188.ani.app.domain.danmaku.DanmakuManager
 import me.him188.ani.app.domain.danmaku.DanmakuManagerImpl
 import me.him188.ani.app.domain.media.cache.DefaultMediaAutoCacheService
