+++ app/shared/ui-settings/src/commonMain/kotlin/ui/settings/framework/AbstractSettingsViewModel.kt
@@ -1,7 +1,16 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
 package me.him188.ani.app.ui.settings.framework
 
 import kotlinx.coroutines.CoroutineScope
-import me.him188.ani.app.data.repository.Settings
+import me.him188.ani.app.data.repository.user.Settings
 import me.him188.ani.app.ui.foundation.AbstractViewModel
 import me.him188.ani.app.ui.foundation.produceState
 import kotlin.properties.PropertyDelegateProvider
