+++ app/shared/ui-settings/src/commonMain/kotlin/ui/settings/framework/SettingsState.kt
@@ -1,3 +1,12 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
 package me.him188.ani.app.ui.settings.framework
 
 import androidx.compose.runtime.Composable
@@ -9,7 +18,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import kotlinx.coroutines.CoroutineScope
-import me.him188.ani.app.data.repository.Settings
+import me.him188.ani.app.data.repository.user.Settings
 import me.him188.ani.app.tools.MonoTasker
 import me.him188.ani.app.ui.foundation.produceState
 import me.him188.ani.utils.platform.annotations.TestOnly
